+++ src/client/rsg-components/SectionHeading/SectionHeading.spec.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { shallow, mount } from 'enzyme';
 import SectionHeading from './index';
 import SectionHeadingRenderer from './SectionHeadingRenderer';
