+++ src/client/rsg-components/SectionHeading/index.ts
